Report on changing Memberships
Add a section to the report on Memberships that have been added or
removed. It's not 100% obvious yet exactly what format will be most
useful for these, so for now try grouping by Term, and displaying only
the person's name and start/end date (if applicable). Once we've seen
it in action for a while it'll hopefully become apparent whether this
is sufficient/suitable.