Make ignored membership fields configurable
Rather than hard-coding the field we want to remove in the method name
it's going to be more future proof to allow the list of fields to
exclude to be configurable. This avoids having to change the method name
in the future when we want to add more fields to be excluded.